Filter useless contexts
Create and use a dedicated ContextSet, instead of BTreeSet, to insert
context entries with precedence checks.  This automatically filter
contexts to ignore or remove useless ones.

This makes it possible to extend test_parse_config_dup_and_sorted() with
a context that doesn't have a when_beneath property, and ignore such
context.

Signed-off-by: Mickaël Salaün <mic@digikod.net>