Create context.when_beneath workspaces
By default, it is useful to be able to modify the projects we are
working on, mainly defined by the context's when_beneath paths.  Make
these paths readable and writable.  We could add a context option to
tune the default access.

New ISLAND_CONTEXT_BENEATH_<index> environment variables contain the
context's when_beneath paths.

Add a reference to the underlying Profile in ResolvedProfile, and update
related code.

Rename Workspace to SymlinkWorkspace and add a new AbsoluteWorkspace.

Signed-off-by: Mickaël Salaün <mic@digikod.net>