Fix workspace race conditions
When several Island processes are launched concurrently and need to
update a profile (e.g., pipe commands in a shell with Island support),
directory and symlink creation might conflict.

Properly handle concurrent profile resolution with a shared lock when
reading and an exclusive lock if changes may be required.

Bump Rust to 1.89 to support File::lock() and File::lock_shared().

Signed-off-by: Mickaël Salaün <mic@digikod.net>