Improve safety of workspace resolution
Improve safety while favoring performance thanks to shared and exclusive
locks for workspace reading or modification.

Create a ProfileLock trait used by the generic resolve_workspaces(),
and implement SharedLock and ExclusiveLock.  This provides strong typing
for workspace read-only or read-write, and avoids the unreachable
io::Error::other handling.

Create a related ProfileGuard struct to properly guard operations that
should be protected by a shared lock or an exclusive lock, and to ensure
unlocking.

Change the resolve_workspaces() return type by moving the NeedsUpdate
case to the error type instead of relying on an Ok(Option).  This makes
it possible to leverage the `?` notation for early returns, simplify
this function.

This is a bit verbose but useful and much safer for future changes.

It would be nice to only open once the profile directory and pass the
file descriptor to both resolve_workspaces() calls and then use this FD
instead of the profile_dir path, but Rust doesn't provide this kind of
functions yet.

Signed-off-by: Mickaël Salaün <mic@digikod.net>