Add integration tests for Island commands
Add a version test to check that the Island command being tested was
build with the same commit as HEAD.

Add tests for the "create" command: one for profile creation wrt to the
current working directory, and another with a specified list of profile
names.

Add tests for the "run" command: implicit and explicit profile(s), exit
code propagation, and restrictions for filesystem and signal.

Signed-off-by: Mickaël Salaün <mic@digikod.net>